refactor: review recommendations
