+++ .husky/pre-commit
@@ -1,3 +1,2 @@
 npm run format:fix
 npm run lint
-npm run test
