+++ src/components/CardList/CardList.module.scss
@@ -9,3 +9,17 @@
 .noResults {
   text-align: center;
 }
+
+.list {
+  width: 100%;
+}
+
+.pagination {
+  margin-top: 20px;
+  display: flex;
+  justify-content: center;
+}
+
+.title {
+  margin: 20px 0;
+}
