+++ src/pages/Main/Main.module.scss
@@ -14,22 +14,8 @@
   align-items: center;
 }
 
-.title {
-  margin-top: 20px;
-}
-
-.pagination {
-  margin-top: 20px;
-  display: flex;
-  justify-content: center;
-}
-
 .content {
   display: flex;
   justify-content: space-between;
   margin-top: 20px;
-}
-
-.list {
-  width: 100%;
 }
