+++ src/utils/helper.ts
@@ -6,4 +6,9 @@ const truncate = (str: string, num: number) => {
   }
 };
 
-export { truncate };
+const range = (start: number, end: number) => {
+  const length = end - start + 1;
+  return Array.from({ length }, (_, index) => index + start);
+};
+
+export { truncate, range };
